// src/logic/data_handling/Enums.ts
export enum pages
{
    // none = 'none',
    home = 'home',
    about = 'about',
    list = 'list',
    circle = 'circle',
    range = 'range',
    grid = 'grid',
}

export enum goals
{
    none = 'none',
    entertain = 'entertain',
    learn = 'learn',
    create = 'create'
}

export enum tools
{
    none = 'none',
    Touchdesigner = 'Touchdesigner',
    DavinciResolve = 'Davinci Resolve',
    Typescript = 'Typescript',
    Python = 'Python',
    Processing = 'Processing',
    Blender = 'Blender',
    Houdini = 'Houdini',
    SuperCollider = 'Super Collider',
    Krita = 'Krita',
    AffinityDesigner = 'Affinity Designer',
    AffinityPhoto = 'Affinity Photo',
    AffinityPublisher = 'Affinity Publisher',
    Twine = 'Twine',
    Phaser = 'Phaser',
    Webpack = 'Webpack',
    RaspberryPI = 'Raspberry PI',
    Arduino = 'Arduino',
    Networking = 'Networking',
    ProjectionMapping = 'Projection mapping',
    Filming = 'Filming',
    Building = 'Building',
    Woodwork = 'Woodwork'
}

export enum themes
{
    none = 'none',
    daily = 'Daily',
    philosophy = 'Philosophy',
    drama = 'Drama',
    interactive = 'Interactive',
    adventure = 'Adventure',
    puzzle = 'Puzzle',
    generative = 'Generative',
    tactile = 'Tactile',
    Analogue = 'Analogue'
}

export enum projectVariables
{
    day = 'day',
    month = 'month',
    year = 'year',

    durationHrs = 'durationHrs',
    teamSize = 'teamSize',
    learnedValue = 'learnedValue',
    endResultValue = 'endResultValue',
}